+++ project_pico_libs/backscatter.h
@@ -63,11 +63,11 @@ uint8_t instructionCount(uint16_t delay, uint16_t max_delay);
 int16_t repeat(uint16_t* instructionBuffer, int16_t delay, uint32_t asm_instr, uint8_t *length, uint16_t max_delay);
 
 // compute the number of delay cycles corresponding to a given phase shift at the center frequency
-uint16_t phase_shift_to_delay_cycles(uint16_t phase_shift, uint16_t d0, uint16_t d1);
+uint32_t phase_shift_to_delay_cycles(uint16_t phase_shift, uint16_t d0, uint16_t d1);
 
 bool generatePIOprogram(uint16_t d0,uint16_t d1, uint32_t baud, uint16_t* instructionBuffer, struct pio_program *backscatter_program);
 
 /* based on d0/d1/baud, the modulation parameters will be computed and returned in the struct backscatter_config */
 void backscatter_program_init(PIO pio, uint8_t *state_machines, uint16_t *pins, uint8_t num_antennae, uint16_t d0, uint16_t d1, uint32_t baud, struct backscatter_config *config, uint16_t *instructionBuffer);
 
-void backscatter_send(PIO pio, uint8_t *state_machines, uint16_t *phase_delay_cycles, uint8_t num_antennae, uint32_t *message, uint32_t len);
+void backscatter_send(PIO pio, uint8_t *state_machines, uint32_t *phase_delay_cycles, uint8_t num_antennae, uint32_t *message, uint32_t len);
